+++ model/image.go
@@ -1,0 +1,5 @@
+package model
+
+type Img struct {
+	//background images coming
+}
