+++ model/tag.go
@@ -1,0 +1,7 @@
+package model
+
+type Tag struct {
+	RuleId int64
+	ImgId  int64
+	Text   string
+}
